Formatting interpolated strings

+++ iDine/models/Order.swift
@@ -11,9 +11,9 @@ import SwiftUI
 class Order: ObservableObject {
     @Published var items = [MenuItem]()
 
-    var total: Int {
+    var total: Double {
         if items.count > 0 {
-            return items.reduce(0) { $0 + $1.price }
+            return Double(items.reduce(0) { $0 + $1.price })
         } else {
             return 0
         }
